requirements.txt
requests~=2.32.3
playwright~=1.47.0
PyYAML~=6.0.2